Refactor cleaner.py to remove unnecessary columns and improve code organization
