+++ steam_sales/steam_etl/sql/get_all_game_data.sql
