+++ steam_sales/steam_etl/sql/get_all_steam_data.sql
