+++ steam_sales/steam_etl/sql/get_all_steamspy_data.sql
