+++ steam_sales/steam_etl/sql/get_new_steam_data.sql
