+++ steam_sales/steam_etl/sql/get_new_steamspy_data.sql
