+++ steam_sales/steam_etl/utils.py
